Major Changes
Added an option to read data from a .csv file. Right now the code is a
little confusing to run, but the next step is to organize the program
into separate blocks to allow an easy experience. Fixed an error with
the freqZoom funtion. Now it should work with any signal provided. Added
a folder to store .csv files with an example.